requirements.txt
Flask==1.1.2
Flask-Session==0.3.2
tcping==0.1.1rc1
mysql-connector-python==8.0.23
mysqlclient==2.0.3
Jinja2==2.11.2
